requirements.txt
Flask
requests
pyjwt
jinja2
pandas
matplotlib
PyYAML
requests
playwright
beautifulsoup4
lxml
pandas
yq
